Añadido tests para clase Personaje

+++ src/clases/personajes/Mortifago.java
@@ -3,7 +3,7 @@ package clases.personajes;
 // ataques oscuros y habilidades especiales
 public class Mortifago extends Personaje {
 
-	public Mortifago(String nombre, double puntosDeVida, int nivelDeMagia, int energia, int multiplicador) {
+	public Mortifago(String nombre, double puntosDeVida, int nivelDeMagia, int energia, double multiplicador) {
 		super(nombre, puntosDeVida, nivelDeMagia, energia, multiplicador);
 	}
